pipa/requirements.txt
python-telegram-bot==20.3
rapidfuzz==2.13.3
python-dotenv
nest_asyncio
